% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-blk-to-stream-name.R
\name{fwa_add_blk_to_stream_name}
\alias{fwa_add_blk_to_stream_name}
\title{Adds Blue Line Key to Stream Name}
\usage{
fwa_add_blk_to_stream_name(x)
}
\arguments{
\item{x}{A data frame with character column StreamName.}
}
\value{
A tibble with the columns of x plus an integer column BLK.
}
\description{
Adds blue line keys (BLK) to stream names.
A stream name may correspond to more than one blue line key (BLK).
}
\examples{
fwa_add_blk_to_stream_name(data.frame(StreamName = "Sangan River"))
}
\seealso{
Other streams: 
\code{\link{fwa_get_stream_name}()}
}
\concept{streams}
