% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-blks-to-stream-name.R
\name{fwa_add_blks_to_stream_name}
\alias{fwa_add_blks_to_stream_name}
\title{Add Blue Line Key(s) to Stream Name}
\usage{
fwa_add_blks_to_stream_name(x)
}
\arguments{
\item{x}{A data frame with character column StreamName.}
}
\value{
A tibble with the columns of x plus an integer column BLK.
}
\description{
Adds blue line keys (BLK) to stream names.
There may be more than one stream with the same name.
}
\examples{
fwa_add_blks_to_stream_name(data.frame(StreamName = "Sangan River"))
}
\seealso{
Other streams: 
\code{\link{fwa_find_stream_names}()}
}
\concept{streams}
