% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-collection-to-watershed.R
\name{fwa_add_collection_to_watershed}
\alias{fwa_add_collection_to_watershed}
\title{Add Collection to Watershed}
\usage{
fwa_add_collection_to_watershed(
  x,
  collection = "stream_network",
  intersect = FALSE,
  filter = NULL,
  limit = 10000,
  offset = 0,
  properties = NULL,
  transform = NULL,
  epsg = getOption("fwa.epsg", 3005),
  camel_case = TRUE
)
}
\arguments{
\item{x}{A sf object with an active sfc polygon column.}

\item{collection}{A character string of the collection.}

\item{intersect}{A logical vector specifying whether to intersect the
individual features with the watershed as opposed to just including
the features that intersect the watershed.}

\item{filter}{A named vector or list of the filter(s) to apply, where the list
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{offset}{A positive whole number indicating the offset of start of returned results.}

\item{properties}{A vector of strings of the column names to include. If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the \href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}} function followed by the parameter values (e.g. c("ST_Simplify", 100))}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{camel_case}{A flag specifying whether to rename collection column names
to upper camel case.}
}
\value{
An sf object
}
\description{
Adds collection to a watershed.
If the active sfc polygon column is called geometry it is replaced
by the geometry column of the collection.
}
\details{
The collection column names are renamed to upper camel case.
T
}
\examples{
\dontrun{
watershed <- fwa_add_watershed_to_blk(data.frame(BLK = 356308001))
fwa_add_collection_to_watershed(watershed)
}
}
\seealso{
\code{\link[fwapgr]{fwa_collection}}.
}
