% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-columns-to-rkm.R
\name{fwa_add_columns_to_rkm}
\alias{fwa_add_columns_to_rkm}
\title{Add Columns to Rkm}
\usage{
fwa_add_columns_to_rkm(rkm, y)
}
\arguments{
\item{rkm}{A data frame with columns blue_line_key and rkm.}

\item{y}{A data frame with columns blue_line_key, rkm and additional columns
to add to rkm.}
}
\value{
An ordered version of rkm with additional columns from y.
}
\description{
Adds columns to rkm based on end river kilometers in y.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}} for \code{\link{fwa_add_end_id_to_rkm}()}
which adds one column at a time.
}
