% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-downstream-split-to-rms.R
\name{fwa_add_downstream_split_to_rms}
\alias{fwa_add_downstream_split_to_rms}
\title{Add Split to River Meter}
\usage{
fwa_add_downstream_split_to_rms(x, y)
}
\arguments{
\item{x}{A data frame with integer columns blk, rm, parent_blk and parent_rm.}

\item{y}{A data frame with integer columns blk and rm and character column name.}
}
\value{
A tibble of x with a logical column for each name in name.
}
\description{
Splits river meters with parent_blk and parent_rm columns
into upstream (TRUE) versus not (FALSE).
}
