% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-end-id-to-rkm.R
\name{fwa_add_end_id_to_rkm}
\alias{fwa_add_end_id_to_rkm}
\title{Add End ID to Rkm}
\usage{
fwa_add_end_id_to_rkm(rkm, y, id = "id")
}
\arguments{
\item{rkm}{A data frame with columns blue_line_key and rkm.}

\item{y}{A data frame with columns blue_line_key, rkm and and id column to add.}

\item{id}{A string of the name of the column in y (can include missing values).}
}
\value{
An ordered version of rkm with id columns from y.
}
\description{
Adds id column to rkm based on end river kilometers in y.
All rkms up to and including the end but before the previous end
are assigned the id (which can be a missing value)
}
