% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-end-id-to-rm.R
\name{fwa_add_end_id_to_rm}
\alias{fwa_add_end_id_to_rm}
\title{Add End ID to River Kilometre}
\usage{
fwa_add_end_id_to_rm(rm, y, id = "id")
}
\arguments{
\item{rm}{A data frame with columns blue_line_key and rm.}

\item{y}{A data frame with columns blue_line_key, rm and and id column to add.}

\item{id}{A string of the name of the column in y (can include missing values).}
}
\value{
An ordered version of rm with id columns from y.
}
\description{
Adds id column to rm based on end river metres in y.
All rms up to and including the end but before the previous end
are assigned the id (which can be a missing value)
}
