% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-fundamental-watershed-to-blk.R
\name{fwa_add_fundamental_watershed_to_blk}
\alias{fwa_add_fundamental_watershed_to_blk}
\title{Add Fundamental Watershed to Blue Line Key}
\usage{
fwa_add_fundamental_watershed_to_blk(x, rm, epsg = getOption("fwa.epsg", 3005))
}
\arguments{
\item{rm}{A positive whole numeric of the distance in meters upstream
from the river mouth.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf tibble with the columns of x plus
sfc polygon column geometry.

A sf object
}
\description{
Adds polygon (geometry) of fundamental watershed to blue line key (BLK).
}
\examples{
\dontrun{
fwa_add_fundamental_watershed_to_blk(data.frame(BLK = 356308001), rm = 100)
}
}
\seealso{
\code{\link[fwapgr]{fwa_watershed_at_measure}}.
}
