% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-intersection-to-geometry.R
\name{fwa_add_intersection_to_geometry}
\alias{fwa_add_intersection_to_geometry}
\title{Add Intersection to Geometry}
\usage{
fwa_add_intersection_to_geometry(x, y)
}
\arguments{
\item{x}{An sf data frame.}

\item{y}{An sf data frame with character column name.}
}
\value{
A copy of x with a logical column for each name in name
indicating whether each element of x intersects with the element of y.
}
\description{
Adds a logical column for each name in name in y
indicating whether each element of x intersects with the element of y.
}
