% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{fwa_add_nearest_id}
\alias{fwa_add_nearest_id}
\title{Add Nearest ID}
\usage{
fwa_add_nearest_id(x, y, id = "id", max_distance = 10)
}
\arguments{
\item{x}{A sf data frame with sfc column and optionally blue_line_key}

\item{y}{A sf data frame with sfc and id column and optionally blue_line_key.}

\item{id}{A string of the name of the positive integer column in y.}

\item{max_distance}{A number of the maximum distance to allow a match.}
}
\value{
An copy of rkm with additional id column from y.
}
\description{
Adds positive integer id to simple features data frame based on
nearest feature in sf data frame and blue_line_key if present.
Rows which are further than the maximum distance from any feature are assigned a missing value.
Useful for assigning line features to rkm table.
Missing values are permitted for the blue_line_key for x but not y.
}
