% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkm.R
\name{fwa_add_nearest_id_to_rkm}
\alias{fwa_add_nearest_id_to_rkm}
\title{Add Nearest ID to Rkm}
\usage{
fwa_add_nearest_id_to_rkm(
  rkm,
  y,
  id = "id",
  max_distance = 10,
  max_end_distance = Inf
)
}
\arguments{
\item{rkm}{A sf data frame with sfc and rkm column and optionally blue_line_key}

\item{y}{A sf data frame with sfc and id column and optionally blue_line_key.}

\item{id}{A string of the name of the positive integer column in y.}

\item{max_distance}{A number of the maximum distance in m to allow a match.}

\item{max_end_distance}{A number of the maximum distance in m to allow a match
for the end rkm for each blue_line_key. Applied recursively to trim the ends.}
}
\value{
An copy of rkm with additional id column from y.
}
\description{
Adds positive integer id to simple features rkm data frame based on
nearest feature in sf data frame and blue_line_key if present.
Rows which are further than the maximum distance from any feature are assigned a missing value.
Missing values are permitted for the blue_line_key for x but not y.
}
