% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-point-to-stream-measure.R
\name{fwa_add_point_to_stream_measure}
\alias{fwa_add_point_to_stream_measure}
\title{Snap Stream Measure to Point}
\usage{
fwa_add_point_to_stream_measure(x, streams, ...)
}
\arguments{
\item{x}{An sf object of spatial points with optional integer column blk.}

\item{streams}{An sf object of spatial linestrings with blk column.}

\item{...}{Additional columns to group by when assigning.}
}
\value{
An updated version of x with integer columns blk and stream_measure and numeric column distance_to_stream.
}
\description{
Assigns closest stream measure in m to each spatial point.
If the blue line key (blk) is missing then it is also assigned
together with the distance to the stream (distance_to_stream) in m.
}
\examples{
\dontrun{
watershed <- fwa_add_watershed_to_blk(data.frame(blk = 356308001, rm = 1000))
network <- fwa_add_collection_to_polygon(watershed)
network$blk <- network$blue_line_key
streams <- fwa_join_stream_segments(network)
points <- fwa_add_rms_to_blk(data.frame(blk = 356308001))
points <- fwa_snap_stream_measure_to_point(points, streams)
points <- points[c("blk", "stream_measure")]
fwa_add_point_to_stream_measure(points, streams)
}
}
\seealso{
\code{\link[=fwa_snap_rm_to_point]{fwa_snap_rm_to_point()}}
}
