% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rm-to-lon-lat.R
\name{fwa_add_rm_to_lon_lat}
\alias{fwa_add_rm_to_lon_lat}
\title{Add River Meters to Longitude and Latitude}
\usage{
fwa_add_rm_to_lon_lat(
  x,
  tolerance = 5000,
  limit = 1,
  epsg = getOption("fwa.epsg", 3005)
)
}
\arguments{
\item{x}{A data frame with numeric longitude (long) and
latitude (lat) columns.}

\item{tolerance}{A number of the tolerance in m.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}

\item{epsg}{A positive whole number of the EPSG projection for the geometry.}
}
\value{
An sf tibble with the columns of x plus integer column blk,
real columns rm and distance_to_lon_lat and sfc point column geometry
}
\description{
Adds numeric river meter (rm) and integer blue line key (blk) column
and sfc point (geometry) column
of the closest point on the stream network (within 5 km)
to the point specified by the lon and lat (WGS84).
}
\details{
If a match isn't found the row is dropped.
}
\examples{
\dontrun{
fwa_add_rm_to_lon_lat(data.frame(lon = -132.26, lat = 53.36))
}
}
