% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rms-to-blk.R
\name{fwa_add_rms_to_blk}
\alias{fwa_add_rms_to_blk}
\title{Add River Meters to Blue Line Key}
\usage{
fwa_add_rms_to_blk(
  x,
  interval = 1000,
  start = 0,
  end = Inf,
  epsg = getOption("fwa.epsg", 3005)
)
}
\arguments{
\item{x}{A data frame with integer column blk.}

\item{interval}{A whole numeric of the distance between points.}

\item{start}{A whole numeric of the start distance.}

\item{end}{An integer of the end distance.}

\item{epsg}{A positive whole number of EPSG projection for the coordinates.}
}
\value{
An sf tibble with the columns of x plus integer column rm
and sf column geometry.
}
\description{
Adds distances (rm) and spatial coordinates (geometry) of
regularly spaced points along blue line key (blk).
All distances which are in meters are from the river mouth.
}
\examples{
fwa_add_rms_to_blk(data.frame(blk = 356308001))
}
\seealso{
Other rm: 
\code{\link{fwa_add_rm_to_lon_lat}()},
\code{\link{fwa_snap_rm_to_point}()}
}
\concept{rm}
