% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-section-to-rm.R
\name{fwa_add_section_to_rm}
\alias{fwa_add_section_to_rm}
\title{Add Section to River Meter}
\usage{
fwa_add_section_to_rm(x, y, section = "Section")
}
\arguments{
\item{x}{A data frame with integer columns BLK and RM.}

\item{y}{A data frame with integer columns BLK and RM and column specified in section.}

\item{section}{A string of the name of the column in y (can include missing values).}
}
\value{
A tibble of x with section column from y.
}
\description{
Adds section column in y to x based on BLK and end RM in y.
All RMs in x up to and including the end RM but before the previous end
RM are assigned the section column value (which can be missing).
}
