% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-section-to-rms.R
\name{fwa_add_section_to_rms}
\alias{fwa_add_section_to_rms}
\title{Add Section to River Meter}
\usage{
fwa_add_section_to_rms(x, y, section = "section")
}
\arguments{
\item{x}{A data frame with integer columns blk and rm.}

\item{y}{A data frame with integer columns blk and rm and column specified in section.}

\item{section}{A string of the name of the column in y (can include missing values).}
}
\value{
A tibble of x with section column from y.
}
\description{
Adds section column in y to x based on blk and end rm in y.
All rms in x up to and including the end rm but before the previous end
rm are assigned the section column value (which can be missing).
}
