% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-watershed-to-blk.R
\name{fwa_add_watershed_to_blk}
\alias{fwa_add_watershed_to_blk}
\title{Add Watershed to Blue Line Key}
\usage{
fwa_add_watershed_to_blk(x, start = 0, epsg = getOption("fwa.epsg", 3005))
}
\arguments{
\item{start}{A positive whole numeric of the distance in meters upstream
from the river mouth.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf tibble with the columns of x plus integer column RM,
numeric column Area and sf column geometry.

A sf object
}
\description{
Adds watershed area (Area) and polygon (geometry) to blue line key (BLK).
The start distances which is in meters is from the river mouth.
}
\examples{
\dontrun{
fwa_blue_line_key_to_watershed(356308001)
}
}
\seealso{
\code{\link[fwapgr]{fwa_watershed_at_measure}}.
}
