% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{fwa_blue_line_key_to_watershed}
\alias{fwa_blue_line_key_to_watershed}
\title{FWA blue_line_key to watershed}
\usage{
fwa_blue_line_key_to_watershed(
  blue_line_key,
  distance_from_mouth = 0,
  epsg = getOption("fwa.epsg", 3005)
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{distance_from_mouth}{A positive whole number of the distance in meters upstream from the river mouth.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Get watershed polygon from blue_line_kay starting from a distance upstream of the river mouth.
}
\examples{
\dontrun{
fwa_blue_line_key_to_watershed(356308001)
}
}
\seealso{
\code{\link[fwapgr]{fwa_watershed_at_measure}}.
}
