% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwa_collection_names}
\alias{fwa_collection_names}
\title{Collection Names}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12 rows and 2 columns.
}
\usage{
fwa_collection_names
}
\description{
A tibble of collection names.
}
\examples{
fwa_collection_names
}
\keyword{datasets}
