% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-rms-to-streams.R
\name{fwa_convert_rms_to_streams}
\alias{fwa_convert_rms_to_streams}
\title{Convert River Metres to Streams}
\usage{
fwa_convert_rms_to_streams(x)
}
\arguments{
\item{x}{An point sf object with whole numeric columns blk and rm.}
}
\value{
A linestring sf object with whole numeric columns blk and rm.
}
\description{
Casts river metre points to linestrings by blk.
}
