% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-stream-network-to-rms.R
\name{fwa_convert_stream_network_to_rms}
\alias{fwa_convert_stream_network_to_rms}
\title{Convert Stream Network to River Meters}
\usage{
fwa_convert_stream_network_to_rms(x, interval = 5, tolerance = 0.1)
}
\arguments{
\item{x}{An sf tibble of a stream network.}

\item{interval}{A whole number of the distance between points.}

\item{tolerance}{A number of the acceptable
discrepancy in meters in the network lengths.}
}
\value{
An sf tibble with the columns of x plus integer column rm
and sf column geometry.
}
\description{
Convert Stream Network to River Meters
}
\examples{
\dontrun{
watershed <- fwa_add_watershed_to_blk(data.frame(blk = 356308001, rm = 1000))
network <- fwa_add_collection_to_watershed(watershed)
fwa_convert_stream_network_to_rms(network, interval = 100)
}
}
\seealso{
Other rm: 
\code{\link{fwa_add_rm_to_lon_lat}()},
\code{\link{fwa_add_rms_to_blk}()},
\code{\link{fwa_snap_rm_to_point}()}
}
\concept{rm}
