% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-streams-to-rms.R
\name{fwa_convert_streams_to_rms}
\alias{fwa_convert_streams_to_rms}
\title{Convert Streams to River Meters}
\usage{
fwa_convert_streams_to_rms(
  x,
  interval = 5,
  gap = 1,
  end = NULL,
  elevation = FALSE,
  reverse = integer()
)
}
\arguments{
\item{x}{An sf tibble with a column blk and linestrings of streams.}

\item{interval}{A positive whole number of the distance (m) between points.}

\item{gap}{A positive real number specifying the maximum gap (m) between
the mouth of stream and its parent stream to be considered connected.}

\item{end}{A positive whole number indicating how far (m) the end of
the stream linestring has to be from the last interval to be included.
To  default \code{end = NULL} (equivalent to \code{end = interval + 1}) excludes ends.}

\item{elevation}{A flag specifying whether to use the elevation
from Google Maps to determine stream direction (or use the
direction of the provided linestrings)}

\item{reverse}{A whole numeric vector of streams to reverse direction
ignoring elevation.}
}
\value{
An sf tibble with the columns blk, integer column rm
and sf column point geometry.
}
\description{
Converts a tibble of streams to river meters.
Unlike \code{\link[=fwa_convert_stream_network_to_rms]{fwa_convert_stream_network_to_rms()}} it only requires
the linestrings and the unique integer identifier for each stream.
}
\examples{
\dontrun{
watershed <- fwa_add_watershed_to_blk(data.frame(blk = 356308001, rm = 1000))
network <- fwa_add_collection_to_polygon(watershed)
network <- select(network, blk = blue_line_key)
fwa_convert_streams_to_rms(network, interval = 100)
}
}
\seealso{
\code{\link[=fwa_convert_stream_network_to_rms]{fwa_convert_stream_network_to_rms()}}
}
