% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-stream-names.R
\name{fwa_find_stream_names}
\alias{fwa_find_stream_names}
\title{Find Stream Names}
\usage{
fwa_find_stream_names(pattern = ".*", ignore_case = TRUE)
}
\arguments{
\item{ignore_case}{A flag specifying whether to ignore case
when matching the regular expression to gnis stream names.}

\item{stream_name}{A string of a regular expression.}
}
\value{
A tibble with character column StreamName of the names of
all the streams that match the regular expression.
}
\description{
Finds gnis stream names that match regular expression.
}
\examples{
fwa_find_stream_names("sangan")
}
\seealso{
Other streams: 
\code{\link{fwa_add_blks_to_stream_name}()}
}
\concept{streams}
