% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{fwa_get_cultural_lines}
\alias{fwa_get_cultural_lines}
\title{FWA get cultural lines}
\usage{
fwa_get_cultural_lines(watershed, limit = getOption("fwa.limit", 1000))
}
\arguments{
\item{watershed}{A single row sf object or sfc object of the watershed polygon (sfc geometry must inherit from sfc_POLYGON).}

\item{limit}{A positive whole number indicating the maximum number of features to return.}
}
\value{
A sf object
}
\description{
Get cultural line features within watershed.
}
\examples{
\dontrun{
watershed <- fwa_blue_line_key_to_watershed(356308001)
fwa_get_cultural_lines(watershed)
}
}
\seealso{
\code{\link[fwapgr]{fwa_collection}}.
}
