% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{fwa_get_manmade_waterbodies}
\alias{fwa_get_manmade_waterbodies}
\title{FWA get manmade waterbodies}
\usage{
fwa_get_manmade_waterbodies(
  watershed,
  intersect = FALSE,
  epsg = getOption("fwa.epsg", 3005),
  limit = getOption("fwa.limit", 10000)
)
}
\arguments{
\item{watershed}{A single row sf object or sfc object of the watershed polygon (sfc geometry must inherit from sfc_POLYGON).}

\item{intersect}{A flag specifying whether to intersect the individual features with the watershed. Useful if the watershed is not a true watershed.}

\item{epsg}{A positive whole number of the epsg to return the collection to.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}
}
\value{
A sf object
}
\description{
Get manmade waterbody features within watershed.
}
\examples{
\dontrun{
watershed <- fwa_blue_line_key_to_watershed(356308001)
fwa_get_manmade_waterbodies(watershed)
}
}
\seealso{
\code{\link[fwapgr]{fwa_collection}}.
}
