% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-section-from-rms.R
\name{fwa_get_section_from_rms}
\alias{fwa_get_section_from_rms}
\title{Get Section from River Meter}
\usage{
fwa_get_section_from_rms(x, section = "section")
}
\arguments{
\item{x}{A data frame with integer columns blk and rm.}

\item{section}{A string of the name of the column in x that specifies the sections.}
}
\value{
A sf tibble with integer columns blk, rm, length and a geometry where rm and geometry are the upstream end of the section.
}
\description{
Gets sf tibble of section blk, rm, length and geometry
where sections are defined by the their most upstream
point.
}
