% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stream-name.R
\name{fwa_get_stream_name}
\alias{fwa_get_stream_name}
\title{Get Stream Name}
\usage{
fwa_get_stream_name(pattern = ".*", ignore_case = TRUE)
}
\arguments{
\item{ignore_case}{A flag specifying whether to ignore case
when matching the regular expression.}

\item{stream_name}{A string of a regular expression.}
}
\value{
A tibble with column StreamName of the names of
all the streams that match the regular expression.
}
\description{
Get gnis stream names that match regular expression.
}
\examples{
fwa_get_stream_name("sangan")
}
\seealso{
Other streams: 
\code{\link{fwa_add_blk_to_stream_name}()}
}
\concept{streams}
