% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-stream-segments.R
\name{fwa_join_stream_segments}
\alias{fwa_join_stream_segments}
\title{Join Stream Segments}
\usage{
fwa_join_stream_segments(x, elevation = FALSE, reverse = integer())
}
\arguments{
\item{x}{An sf tibble with a column blk and linestrings of stream segments.}

\item{elevation}{A flag specifying whether to use the elevation
from Google Maps to determine stream direction (or use the
direction of the provided linestrings)}

\item{reverse}{A whole numeric vector of streams to reverse direction
ignoring elevation.}
}
\value{
An sf tibble with the columns blk and sfc column point geometry.
}
\description{
Converts a tibble of stream segment linestrings to stream linestrings.
}
\examples{
\dontrun{
watershed <- fwa_add_watershed_to_blk(data.frame(blk = 356308001, rm = 1000))
network <- fwa_add_collection_to_polygon(watershed)
network <- select(network, blk = blue_line_key)
fwa_join_stream_segments(network)
}
}
\seealso{
\code{\link[=fwa_convert_streams_to_rms]{fwa_convert_streams_to_rms()}}
}
