% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview.R
\name{fwa_mapview}
\alias{fwa_mapview}
\title{Map View}
\usage{
fwa_mapview(x, layer = NULL, zcol = NULL, legend = FALSE, ...)
}
\arguments{
\item{x}{An sf data frame.}

\item{layer}{A string of the column to layer the points by.}

\item{zcol}{A string of the column to color points by.}

\item{legend}{A flag specifying whether to plot a legend.}

\item{...}{Additional arguments passed to \code{mapview::mapview()}.}
}
\description{
A wrapper on \code{mapview::mapview.sf()} that allows the user to layer by a column
and coerces hms columns to character (to avoid being dropped).
}
