% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-rms.R
\name{fwa_mapview_rms}
\alias{fwa_mapview_rms}
\title{Map View River Meters}
\usage{
fwa_mapview_rms(x, zcol = "rm", npoint = 500)
}
\arguments{
\item{x}{An sf data frame with unique integer columns blk and rm.}

\item{zcol}{A string of the column to color points by.}

\item{npoint}{An indication of the total number of points to plot.}
}
\description{
Map View River Meters
}
