% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkm.R
\name{fwa_nearest_rkm}
\alias{fwa_nearest_rkm}
\title{Nearest Rkm}
\usage{
fwa_nearest_rkm(x, rkm)
}
\arguments{
\item{x}{An sf data frame of spatial points with optional blue_line_key column.}

\item{rkm}{An sf data frame of spatial point with blue_line_key and rkm columns.}
}
\value{
An updated version of x with column rkm and distance_to_rkm in m.
}
\description{
Assigns spatial points to river kilometre with matching blue_line_key.
If the blue_line_key is missing then it also assigns the blue_line_key.
}
\examples{
\dontrun{
rkm <- fwa_rkm(blue_line_key = 356308001, interval = 1000)
fwa_nearest_rkm(rkm[2:4,], rkm[3:5,])
}
}
\seealso{
Other rkm: 
\code{\link{fwa_rkm}()}
}
\concept{rkm}
