% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest-rm.R
\name{fwa_nearest_rm}
\alias{fwa_nearest_rm}
\title{Nearest River Metre}
\usage{
fwa_nearest_rm(x, rm)
}
\arguments{
\item{x}{An sf data frame of spatial points with optional blue_line_key column.}

\item{rm}{An sf data frame of spatial point with blue_line_key and rm columns.}
}
\value{
An updated version of x with column rm and distance_to_rm in m.
}
\description{
Assigns spatial points to river metre with matching blue_line_key.
If the blue_line_key is missing then it also assigns the blue_line_key.
}
\examples{
\dontrun{
rm <- fwa_rm(blue_line_key = 356308001, interval = 1000)
fwa_nearest_rm(rm[2:4,], rm[3:5,])
}
}
\seealso{
Other rm: 
\code{\link{fwa_rm}()}
}
\concept{rm}
