% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent-stream-name-rms.R
\name{fwa_parent_stream_name_rms}
\alias{fwa_parent_stream_name_rms}
\title{Parent Stream Name}
\usage{
fwa_parent_stream_name_rms(x, rms, stream_name = fwatlasbc::fwa_stream_name)
}
\arguments{
\item{x}{A character vector of one or more stream names.}

\item{rms}{A data frame with integer columns blk and parent_blk.
There must be only one parent_blk for each blk.}

\item{stream_name}{A data frame with character column stream_name and integer column blk. There must be no more than one blk for the stream names specified.}
}
\value{
A character vector of the parent blue line keys.
}
\description{
Gets parent stream name.
}
