% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{fwa_point_to_watershed}
\alias{fwa_point_to_watershed}
\title{FWA point to watershed}
\usage{
fwa_point_to_watershed(point)
}
\arguments{
\item{point}{A single row sf object or sfc object of the point (sfc geometry must inherit from sfc_POINT).}
}
\value{
A sf object
}
\description{
Get watershed polygon from a sf or sfc point starting from nearest point along a FWA stream.
}
\examples{
\dontrun{
sf_point <- sf::st_as_sf(sf::st_as_sfc("POINT(-131.927 54.032)", crs = 4326))
fwa_point_to_watershed(sf_point)
}
}
\seealso{
\code{\link[fwapgr]{fwa_watershed_at_measure}} and \code{\link[fwapgr]{fwa_index_point}}.
}
