% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune-rms.R
\name{fwa_prune_rms}
\alias{fwa_prune_rms}
\title{Prune River Meter}
\usage{
fwa_prune_rms(x, y)
}
\arguments{
\item{x}{A data frame with integer columns blk, rm, parent_blk and parent_rm.}

\item{y}{A data frame with integer columns blk and rm.}
}
\value{
A tibble of x with rows above the rms in y removed.
}
\description{
Removes river meters above rm values in y.
}
