% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkm.R
\name{fwa_rkm}
\alias{fwa_rkm}
\title{Get Rkm}
\usage{
fwa_rkm(
  blue_line_key,
  interval = 1000,
  start = 0,
  end = NULL,
  epsg = getOption("fwa.epsg", 3005)
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{interval}{An integer of the interval distance in meters.}

\item{start}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{end}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Provided a blue_line_key and start and possibly end distance upstream from
river mouth, return river kilometre points at regular intervals.
}
\examples{
fwa_rkm(356308001)
}
\seealso{
Other rkm: 
\code{\link{fwa_nearest_rkm}()}
}
\concept{rkm}
