% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{fwa_rm}
\alias{fwa_rm}
\title{Get River Metre}
\usage{
fwa_rm(
  blue_line_key,
  interval = 1000,
  start = 0,
  end = NULL,
  epsg = getOption("fwa.epsg", 3005),
  limit = getOption("fwa.limit", 10000)
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{interval}{An integer of the interval distance in meters.}

\item{start}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{end}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{limit}{A positive whole number of the maximum number of features to return.}
}
\value{
A sf object
}
\description{
Provided a blue_line_key and start and possibly end distance upstream from
river mouth, return river metre points at regular intervals.
}
\examples{
fwa_rm(356308001)
}
\seealso{
Other rm: 
\code{\link{fwa_nearest_rm}()}
}
\concept{rm}
