% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-rm-to-point.R
\name{fwa_snap_rm_to_point}
\alias{fwa_snap_rm_to_point}
\title{Snap River Meter to Point}
\usage{
fwa_snap_rm_to_point(x, rm)
}
\arguments{
\item{x}{An sf object of spatial points with optional integer column BLK.}

\item{rm}{An sf object of spatial point with BLK and RM columns.}
}
\value{
An updated version of x with integer columns BLK and RM and numeric column Distance.
}
\description{
Assigns closest river meter to each spatial point.
If the blue line key (BLK) is missing then it is also assigned
together with the distance to the river meter (DistanceToRM) in m.
}
\examples{
\dontrun{
rm <- fwa_rm(blue_line_key = 356308001, interval = 1000)
fwa_snap_rm_to_point(rm[2:4,], rm[3:5,])
}
}
\seealso{
Other rm: 
\code{\link{fwa_add_rms_to_blk}()}
}
\concept{rm}
