% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-rm-to-rms.R
\name{fwa_snap_rm_to_rms}
\alias{fwa_snap_rm_to_rms}
\title{Snap River Meter to River Meters}
\usage{
fwa_snap_rm_to_rms(x, rm)
}
\arguments{
\item{x}{An sf object of spatial points with blk and rm columns.}

\item{rm}{An sf object of spatial point with blk and rm columns.}
}
\value{
An updated version of x with integer columns blk, rm and new_rm and numeric column distance_to_new_rm.
}
\description{
Assigns closest river meter to river meters based on blue line keys.
}
\examples{
rm <- fwa_add_rms_to_blk(data.frame(blk = 356308001))
x <- rm[rm$rm \%in\% c(0, 2000, 5000, 6000, 7000),]
rm <- rm[rm$rm \%in\% c(1000, 3000, 4000, 8000, 9000, 10000),]
fwa_snap_rm_to_rms(x, rm)
}
