% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-rms-to-rms.R
\name{fwa_snap_rms_to_rms}
\alias{fwa_snap_rms_to_rms}
\title{Snap River Meters to River Meters}
\usage{
fwa_snap_rms_to_rms(x, rm, snap_mouths = FALSE)
}
\arguments{
\item{x}{An sf object of spatial points with blk and rm columns and
optional new_blk and new_rm integer column.}

\item{rm}{An sf object of spatial point with blk and rm columns.}

\item{snap_mouths}{A flag specifying whether to snap pairs of streams at their mouths (rm = 0) where new_rm is not already set.}
}
\value{
A named list with an updated versions of x and rm with
integer columns blk, new_blk, rm and new_rm and
numeric column distance_to_new_rm.
}
\description{
Assigns closest river meters to river meters by blue line keys using
\code{fwa_snap_rm_to_rms()}
rm must not have an existing new_rm column.
}
\details{
x is first snapped to rm then rm is snapped to x while ensuring
that the links between x and rm are bidirectional as much as possible.
}
\examples{
rm <- fwa_add_rms_to_blk(data.frame(blk = 356308001))
x <- rm[rm$rm \%in\% c(0, 2000, 5000, 6000, 7000), ]
rm <- rm[rm$rm \%in\% c(1000, 3000, 4000, 8000, 9000, 10000), ]
fwa_snap_rms_to_rms(x, rm)
}
\seealso{
\code{\link[=fwa_snap_rm_to_rms]{fwa_snap_rm_to_rms()}}
}
