% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwa_stream_name}
\alias{fwa_stream_name}
\title{Stream Name}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11610 rows and 2 columns.
}
\usage{
fwa_stream_name
}
\description{
A tibble of stream names with blue line keys.
}
\examples{
fwa_stream_name
}
\keyword{datasets}
