% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{fwa_stream_to_blue_line_keys}
\alias{fwa_stream_to_blue_line_keys}
\title{Blue Line Key from Stream}
\usage{
fwa_stream_to_blue_line_keys(stream_name)
}
\arguments{
\item{stream_name}{A string of the stream name.}
}
\value{
A integer vector of blue line keys associated with the stream name.
If no associated blue line keys it returns an empty integer vector.
}
\description{
Gets blue line key from stream name.
}
\examples{
fwa_stream_to_blue_line_keys("Sangan River")
}
