% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams-to-blue-line-keys.R
\name{fwa_streams_to_blue_line_keys}
\alias{fwa_streams_to_blue_line_keys}
\title{Blue Line Keys from Stream Names}
\usage{
fwa_streams_to_blue_line_keys(stream_name)
}
\arguments{
\item{stream_name}{A character vector of stream names.}
}
\value{
A tibble of the stream name and blue line key.
}
\description{
Gets blue line keys from stream names.
}
\examples{
fwa_streams_to_blue_line_keys("Sangan River")
}
\seealso{
Other streams: 
\code{\link{fwa_add_blk_to_stream_name}()},
\code{\link{fwa_search_streams}()}
}
\concept{streams}
