% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{fwa_xy_to_watershed}
\alias{fwa_xy_to_watershed}
\title{FWA xy to watershed}
\usage{
fwa_xy_to_watershed(x, y, epsg = getOption("fwa.epsg", 3005))
}
\arguments{
\item{x}{A number of the x coordinate.}

\item{y}{A number of the y coordinate.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
A sf object
}
\description{
Get watershed polygon from x coordinate, y coordinate and epsg starting from nearest point along a FWA stream.
}
\examples{
\dontrun{
sf_point <- sf::st_as_sf(sf::st_as_sfc("POINT(-131.927 54.032)", crs = 4326))
fwa_xy_to_watershed(x = -131.927, y = 54.032, epsg = 4326)
}
}
\seealso{
\code{\link[fwapgr]{fwa_watershed_at_measure}} and \code{\link[fwapgr]{fwa_indexpoint}}.
}
