% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwatlasbc-package.R
\docType{package}
\name{fwatlasbc-package}
\alias{fwatlasbc}
\alias{fwatlasbc-package}
\title{fwatlasbc: Freshwater Atlas British Columbia}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Sebastian Dalgarno \email{dalgarno@unbc.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
