% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{named_streams}
\alias{named_streams}
\title{Named Streams}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10000 rows and 6 columns.
}
\usage{
named_streams
}
\description{
A tibble of the named streams
}
\keyword{datasets}
