% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-blk-to-stream-name.R
\name{fwa_add_blk_to_stream_name}
\alias{fwa_add_blk_to_stream_name}
\title{Adds Blue Line Key to Stream Name}
\usage{
fwa_add_blk_to_stream_name(x)
}
\arguments{
\item{x}{A data frame with column stream_name.}
}
\value{
An copy of x with additional blue_line_key column.
}
\description{
Adds blue line keys to stream names.
}
\examples{
fwa_add_blk_to_stream_name(data.frame(stream_name = "Sangan River"))
}
\seealso{
Other streams: 
\code{\link{fwa_search_streams}()},
\code{\link{fwa_streams_to_blue_line_keys}()}
}
\concept{streams}
