% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-gm-elevation-to-point.R
\name{fwa_add_gm_elevation_to_point}
\alias{fwa_add_gm_elevation_to_point}
\title{Add Google Maps Elevation to Point}
\usage{
fwa_add_gm_elevation_to_point(
  x,
  key = Sys.getenv("GOOGLE_MAPS_ELEVATION_API_KEY")
)
}
\arguments{
\item{x}{An sf object of spatial points.}

\item{key}{A string of the google maps elevation api key.}
}
\value{
An updated version of x with numeric column elevation.
}
\description{
Add Google Maps Elevation to Point
}
\examples{
\dontrun{
rm <- fwa_add_rms_to_blk(data.frame(blk = 356308001))
fwa_add_gm_elevation_to_point(rm)
}
}
\seealso{
\code{googleway::google_elevation()}
}
