% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkm.R
\name{fwa_add_nearest_id_to_rkm}
\alias{fwa_add_nearest_id_to_rkm}
\title{Add Nearest ID to Rkm}
\usage{
fwa_add_nearest_id_to_rkm(rkm, y, id = "id", max_distance = 10)
}
\arguments{
\item{rkm}{A sf data frame with sfc and rkm column and optionally blue_line_key}

\item{y}{A sf data frame with sfc and id column and optionally blue_line_key.}

\item{id}{A string of the name of the positive integer column in y.}

\item{max_distance}{A number of the maximum distance to allow a match.}
}
\value{
An copy of rkm with additional id column from y.
}
\description{
Adds positive integer id to simple features rkm data frame based on
nearest feature in sf data frame and blue_line_key if present.
Rows which are further than the maximum distance from any feature are assigned a missing value.
Missing values are permitted for the blue_line_key for x but not y.
}
