% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-new-blk-rm-to-blk-rm.R
\name{fwa_add_new_blk_rm_to_blk_rm}
\alias{fwa_add_new_blk_rm_to_blk_rm}
\title{Add New Blue Line Key and River Meter to Blue Line Key and River Meter}
\usage{
fwa_add_new_blk_rm_to_blk_rm(x, rm)
}
\arguments{
\item{x}{A tibble (or sf object) with blk and rm columns.}

\item{rm}{A tibble (or sf object) with unique blk and rm columns
and new blk and rm columns.}
}
\value{
An updated version of x with
additional new_blk and new_rm columns.
}
\description{
Adds new blk and rm values where existing blk and rm values
based on look up in rm.
}
