% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-stream-network-to-blk.R
\name{fwa_add_stream_network_to_blk}
\alias{fwa_add_stream_network_to_blk}
\title{Add Stream Network to Blue Line Key}
\usage{
fwa_add_stream_network_to_blk(x, epsg = getOption("fwa.epsg", 3005))
}
\arguments{
\item{x}{A data frame with a whole numeric blue line key (blk) column
and an optional rm column specfying the river meter.
The rm is set to be 0 if missing.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf object
}
\description{
Adds stream network to blue line key upstream of rm based on watershed codes
and intersection of stream segment start points with blue line key.
}
\examples{
\dontrun{
fwa_add_stream_network_to_blk(data.frame(blk = 356308001))
}
}
\seealso{
\code{\link[fwapgr]{fwa_collection}}.
}
