% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-watershed-to-blk.R
\name{fwa_add_watershed_to_blk}
\alias{fwa_add_watershed_to_blk}
\title{Add Watershed to Blue Line Key}
\usage{
fwa_add_watershed_to_blk(
  x,
  rm = 0,
  exclude = FALSE,
  epsg = getOption("fwa.epsg", 3005)
)
}
\arguments{
\item{rm}{A positive whole numeric of the distance in meters upstream
from the river mouth.}

\item{exclude}{A logical vector specifying whether to exclude the
fundamental watershed in which the start falls.}

\item{epsg}{A positive whole number of the epsg to transform features to.}
}
\value{
An sf tibble with the columns of x plus
sf column geometry.

A sf object
}
\description{
Adds polygon (geometry) of aggregated fundamental watersheds to blue line key (BLK).
The rm distances which is in meters is from the river mouth.
}
\examples{
\dontrun{
fwa_add_watershed_to_blk(data.frame(BLK = 356308001))
}
}
\seealso{
\code{\link[fwapgr]{fwa_watershed_at_measure}}.
}
