% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{fwa_get}
\alias{fwa_get}
\title{FWA get collection}
\usage{
fwa_get(watershed, collection_id, limit = getOption("fwa.limit", 1000))
}
\arguments{
\item{watershed}{A single row sf object or sfc object of the watershed polygon (sfc geometry must inherit from sfc_POLYGON).}

\item{collection_id}{A character string of the collection id.
names correspond to column names and the list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{limit}{A positive whole number indicating the maximum number of features to return.}
}
\value{
A sf object
}
\description{
Get collection within watershed.
}
\examples{
\dontrun{
watershed <- fwa_blue_line_key_to_watershed(356308001)
fwa_get(watershed, "whse_basemapping.fwa_stream_networks_sp"))
}
}
\seealso{
\code{\link[fwapgr]{fwa_collection}}.
}
