% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview-rms-to-rms.R
\name{fwa_mapview_rms_to_rms}
\alias{fwa_mapview_rms_to_rms}
\title{Map View River Meters}
\usage{
fwa_mapview_rms_to_rms(x, y, zcol = "rm", npoint = 250)
}
\arguments{
\item{x}{An sf data frame with unique integer columns blk and rm and integer column new_rm.}

\item{y}{An sf data frame with unique integer columns blk and rm.}

\item{zcol}{A string of the column to color points by.}

\item{npoint}{An indication of the total number of points to plot.}
}
\description{
Maps two alternative stream networks by adding links from each point in x to matching point in y.
}
