% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkm.R
\name{fwa_rkm}
\alias{fwa_rkm}
\title{FWA rkm}
\usage{
fwa_rkm(
  blue_line_key,
  interval = 1000,
  distance_upstream = 0,
  epsg = 4326,
  limit = 500
)
}
\arguments{
\item{blue_line_key}{An integer of the stream blue line key.}

\item{interval}{An integer of the interval distance in meters.}

\item{distance_upstream}{An integer of the distance in meters upstream from the river mouth to start from.}

\item{epsg}{A positive whole number of the epsg to transform features to.}

\item{limit}{A positive whole number of the maximum number of features to return.}
}
\value{
A sf object
}
\description{
Provided a blue_line_key and starting distance upstream from river mouth, return Rkm points along an interval.
}
\examples{
fwa_rkm(356308001)
}
