% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-rm-to-point.R
\name{fwa_snap_rm_to_point}
\alias{fwa_snap_rm_to_point}
\title{Snap River Meter to Point}
\usage{
fwa_snap_rm_to_point(x, rm, ...)
}
\arguments{
\item{x}{An sf object of spatial points with optional integer column blk.}

\item{rm}{An sf object of spatial point with blk and rm columns.}

\item{...}{Additional columns to group by when assigning.
Elements with missing values in rm are assigned to any value in x.}
}
\value{
An updated version of x with integer columns blk and rm and numeric column distance_to_rm.
}
\description{
Assigns closest river meter to each spatial point.
If the blue line key (blk) is missing then it is also assigned
together with the distance to the river meter (distance_to_rm) in m.
}
\examples{
rm <- fwa_add_rms_to_blk(data.frame(blk = 356308001))
x <- rm[rm$rm \%in\% c(0, 2000, 5000, 6000, 7000),]
rm <- rm[rm$rm \%in\% c(1000, 3000, 4000, 8000, 9000, 10000),]
fwa_snap_rm_to_point(x, rm)
}
