% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-stream-measure-to-point.R
\name{fwa_snap_stream_measure_to_point}
\alias{fwa_snap_stream_measure_to_point}
\title{Snap Stream Measure to Point}
\usage{
fwa_snap_stream_measure_to_point(x, streams, ...)
}
\arguments{
\item{x}{An sf object of spatial points with optional integer column blk.}

\item{streams}{An sf object of spatial linestrings with blk column.}

\item{...}{Additional columns to group by when assigning.}
}
\value{
An updated version of x with integer columns blk and stream_measure and numeric column distance_to_stream.
}
\description{
Assigns closest stream measure in m to each spatial point.
If the blue line key (blk) is missing then it is also assigned
together with the distance to the stream (distance_to_stream) in m.
}
\examples{
\dontrun{
watershed <- fwa_add_watershed_to_blk(data.frame(blk = 356308001, rm = 1000))
network <- fwa_add_collection_to_polygon(watershed)
network <- select(network, blk = blue_line_key)
points <- fwa_convert_streams_to_rms(streams, interval = 500) |>
fwa_snap_stream_measure_to_point(points)
}
}
\seealso{
\code{\link[=fwa_snap_rm_to_point]{fwa_snap_rm_to_point()}}
}
