% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{fwa_streams_to_blue_line_keys}
\alias{fwa_streams_to_blue_line_keys}
\title{Blue Line Key from Stream}
\usage{
fwa_streams_to_blue_line_keys(stream_name)
}
\arguments{
\item{stream_name}{A vector of character strings of the stream names.}
}
\value{
A data.frame including blue_line_key and watershed_group_code
values associated with each stream name.
If no associated blue line keys it returns an empty integer vector.
}
\description{
Gets blue line key from stream name.
}
\examples{
fwa_streams_to_blue_line_keys("Sangan River")
}
